Represent signed and unsigned as the same type in Java to avoid wasting space on arrays.
Note that extra care must be taken when converting unsigned values on the Java side.

+++ binary/generate/java.go
@@ -144,11 +144,11 @@ class ObjectFactory {
 	javaFile      *template.Template
 	javaTypeMap   = map[string]string{
 		"int8":    "byte",
-		"uint8":   "short",
+		"uint8":   "byte",
 		"int16":   "short",
-		"uint16":  "int",
+		"uint16":  "short",
 		"int32":   "int",
-		"uint32":  "long",
+		"uint32":  "int",
 		"int64":   "long",
 		"uint64":  "long",
 		"float32": "float",
